% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{olym_final}
\alias{olym_final}
\title{olym_final: Dataset for Olympic Medals Analysis}
\format{
A data frame with several columns, including:
\describe{
  \item{Name}{Athlete's name}
  \item{Sex}{Athlete's Gender}
  \item{Age}{Athlete's age}
  \item{Height}{Athlete's Gender}
  \item{Weight}{Athlete's Gender}
  \item{Team}{The country where the athlete's originated}
  \item{NOC}{Abbreviation of Olympic Entity Represented}
  \item{Year}{The year of Olympics participation}
  \item{City}{The city as the Olympics' Venue}
  \item{Sport}{The sport type participated}
  \item{Event}{The sport number or event participated}
  \item{Year}{The year of Olympics participation}
  \item{Top_1_attr}{The main attribute of the sport}
  \item{Top_2_attr}{The second highest attribute of the sport}
  \item{Top_3_attr}{The third highest attribute of the sport}
  \item{Top_1_value}{The numerical value of the main attribute}
  \item{Top_2_value}{The numerical value of second highest attribute}
  \item{Top_1_value}{The numerical value of third highest attribute}

}
}
\source{
Cleaned data from Olympic datasets.
}
\usage{
olym_final
}
\description{
This dataset contains information about Olympic medals and attributes.
}
\keyword{datasets}
